benchmarks/requirements.txt
tiktoken==0.12.0
pyyaml==6.0.3
python-dotenv==1.2.2

# LLM SDKs (optional - only needed for benchmark_**.py)
openai==2.32.0
anthropic==0.95.0
# TOON comparison (optional - only needed for benchmark_toon_comparison.py)
toon-py==1.0.2